// src/types.ts
export interface HourlyForecastItem {
  time: string;
  temperature: number;
  precipitationProbability: number;
  precipitation: number;
  weatherCode: number;
  weatherDescription: string;
  windSpeed: number;
  relativeHumidity: number;
  uvIndex: number;
}

export interface DailyForecastItem {
  date: string;
  dayName: string;
  tempMax: number;
  tempMin: number;
  apparentTempMax: number;
  apparentTempMin: number;
  precipitationSum: number;
  precipitationProbabilityMax: number;
  weatherCode: number;
  weatherDescription: string;
  windSpeedMax: number;
  uvIndexMax: number;
  sunrise: string;
  sunset: string;
}